requirements.txt
# Discord
discord.py==0.16.7
aiohttp==1.0.5
async-timeout==1.2.0
chardet==2.3.0
websockets==3.2
multidict==2.1.4

# Matplotlib
cycler==0.10.0
matplotlib==2.0.0
numpy==1.12.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0

# database
SQLAlchemy==1.1.6
